Update Exception handlers in Service and Repository layer

+++ simple-expenses/src/main/java/sg/edu/ntu/simpleexpenses/service/ExpensesService.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import sg.edu.ntu.simpleexpenses.exception.OperationFailedException;
+import sg.edu.ntu.simpleexpenses.exception.ResourceNotFoundException;
 import sg.edu.ntu.simpleexpenses.pojo.ExpenseCategory;
 import sg.edu.ntu.simpleexpenses.pojo.Expenses;
 import sg.edu.ntu.simpleexpenses.repository.ExpensesRepository;
@@ -23,6 +25,9 @@ public class ExpensesService {
      * GET ALL
      */
     public List<Expenses> getAllExpenses() {
+        if (expensesRepository.getAllExpenses().isEmpty()) {
+            throw new ResourceNotFoundException("expenses not found");
+        }
         return expensesRepository.getAllExpenses();
     }
 
@@ -30,13 +35,19 @@ public class ExpensesService {
      * GET ONE
      */
     public Expenses getExpense(String id) {
+        if (expensesRepository.getExpense(id) == null) {
+            throw new ResourceNotFoundException("expense not found for ID: " + id);
+        }
         return expensesRepository.getExpense(id);
     }
 
     /*
-     * GET EXPENSES BY CATEGORY
+     * GET EXPENSES BY CATEGORY, MIN AMT, MAX AMT
      */
     public List<Expenses> getExpensesByCategory(ExpenseCategory category, Double minAmount, Double maxAmount) {
+        if (expensesRepository.getExpensesByCategory(category, minAmount, maxAmount).isEmpty()) {
+            throw new ResourceNotFoundException("expense not found");
+        }
         return expensesRepository.getExpensesByCategory(category, minAmount, maxAmount);
     }
 
@@ -44,6 +55,9 @@ public class ExpensesService {
      * CREATE
      */
     public Expenses addExpense(Expenses newExpense) {
+        if (expensesRepository.addExpense(newExpense) == null) {
+            throw new OperationFailedException("failed to add expense");
+        }
         return expensesRepository.addExpense(newExpense);
     }
 
@@ -51,13 +65,19 @@ public class ExpensesService {
      * UPDATE
      */
     public Expenses updateExpense(Expenses updateExpense) {
+        if (expensesRepository.updateExpense(updateExpense) == null) {
+            throw new OperationFailedException("failed to update expense");
+        }
         return expensesRepository.updateExpense(updateExpense);
     }
 
     /*
      * DELETE
      */
-    public List<Expenses> deleteExpense(String id) {
+    public Expenses deleteExpense(String id) {
+        if (expensesRepository.deleteExpense(id) == null) {
+            throw new ResourceNotFoundException("Expense with ID " + id + " not found.");
+        }
         return expensesRepository.deleteExpense(id);
     }
